split ziti-edge-tunnel monolith
