+++ programs/ziti-edge-tunnel/CMakeLists.txt
@@ -23,11 +23,15 @@ set(ZITI_INSTANCE_COMMON
         include/model/events.h
         include/model/dtos.h
         instance.c
+        model.c
         include/identity-utils.h
         config-utils.c
         include/config-utils.h
         instance-config.c
         include/instance-config.h
+        process_cmd.c
+        ipc_cmd.c
+        ipc_event.c
 )
 if (WIN32)
     set(ZITI_INSTANCE_OS
