+++ programs/ziti-edge-tunnel/ziti-edge-tunnel.c
@@ -72,9 +72,7 @@ char* get_log_file_name();
 static int dns_miss_status = DNS_REFUSE;
 
 typedef char * (*to_json_fn)(const void * msg, int flags, size_t *len);
-static void send_events_message(const void *message, to_json_fn to_json_f, bool displayEvent);
-static void send_tunnel_command(const tunnel_command *tnl_cmd, void *ctx);
-static void send_tunnel_command_inline(const tunnel_command *tnl_cmd, void *ctx);
+void send_tunnel_command_inline(const tunnel_command *tnl_cmd, void *ctx);
 static void scm_service_stop_event(uv_loop_t *loop, void *arg);
 static void stop_tunnel_and_cleanup();
 static bool is_host_only();
@@ -115,35 +113,18 @@ struct cfg_instance_s {
 // temporary list to pass info between parse and run
 static LIST_HEAD(instance_list, cfg_instance_s) load_list;
 
-struct event_conn_s {
-    uv_pipe_t *event_client_conn;
-    LIST_ENTRY(event_conn_s) _next_event;
-};
-// list to store the event connections
-static LIST_HEAD(events_list, event_conn_s) event_clients_list = LIST_HEAD_INITIALIZER(event_clients_list);
-
-struct ipc_conn_s {
-    uv_pipe_t ipc;
-    LIST_ENTRY(ipc_conn_s) _next_ipc_cmd;
-};
-// list to store the ipc connections
-static LIST_HEAD(ipc_list, ipc_conn_s) ipc_clients_list = LIST_HEAD_INITIALIZER(ipc_clients_list);
-
 static long refresh_metrics = 5000;
 static long metrics_latency = 5000;
 static char *configured_cidr = NULL;
 static char *configured_log_level = NULL;
 static char *configured_proxy = NULL;
-static char *config_dir = NULL;
-
-static uv_pipe_t cmd_server;
-static uv_pipe_t event_server;
+char *config_dir = NULL;
 
 //timer
 static uv_timer_t metrics_timer;
 
 // singleton
-static const ziti_tunnel_ctrl *CMD_CTRL;
+const ziti_tunnel_ctrl *CMD_CTRL;
 
 static bool started_by_scm = false;
 static bool tunnel_interrupted = false;
@@ -165,737 +146,16 @@ static char sockfile[] = SOCKET_PATH "/ziti-edge-tunnel.sock";
 static char eventsockfile[] = SOCKET_PATH "/ziti-edge-tunnel-event.sock";
 #endif
 
-static int sizeof_event_clients_list() {
-    struct event_conn_s *event_client;
-    int size = 0;
-    LIST_FOREACH(event_client, &event_clients_list, _next_event) {
-        size++;
-    }
+extern int start_cmd_socket(uv_loop_t *l, const char *sockfile);
+extern int start_event_socket(uv_loop_t *l, const char *eventsockfile);
 
-    if (size == 0) {
-        return size;
-    }
-
-    int current_size = size;
-
-    // clean up closed event connection from the list
-    for (int idx = 0; idx < size; idx++) {
-        struct event_conn_s *del_event_client = NULL;
-        LIST_FOREACH(del_event_client, &event_clients_list, _next_event) {
-            if (del_event_client->event_client_conn == NULL) {
-                break;
-            }
-        }
-        if (del_event_client) {
-            LIST_REMOVE(del_event_client, _next_event);
-            free(del_event_client);
-            current_size--;
-        } else {
-            // break from for loop
-            break;
-        }
-    }
-
-    return current_size;
-
-}
-
-static int sizeof_ipc_clients_list() {
-    struct ipc_conn_s *ipc_client;
-    int size = 0;
-    LIST_FOREACH(ipc_client, &ipc_clients_list, _next_ipc_cmd) {
-        size++;
-    }
-    return size;
-}
-
-static void cmd_alloc(uv_handle_t *s, size_t sugg, uv_buf_t *b) {
-    b->base = malloc(sugg);
-    b->len = sugg;
-}
-
-static void send_tunnel_status(char* status) {
+void send_tunnel_status(char* status) {
     tunnel_status_event tnl_sts_evt = {0};
     tnl_sts_evt.Op = strdup(status);
     tnl_sts_evt.Status = get_tunnel_status();
     send_events_message(&tnl_sts_evt, (to_json_fn) tunnel_status_event_to_json, true);
     tnl_sts_evt.Status = NULL; //don't free
     free_tunnel_status_event(&tnl_sts_evt);
-}
-
-static void on_cmd_write(uv_write_t *wr, int len) {
-    if (wr->data) {
-        free(wr->data);
-    }
-    free(wr);
-}
-
-static void on_command_resp(const tunnel_result* result, void *ctx) {
-    size_t json_len;
-    char *json = tunnel_result_to_json(result, MODEL_JSON_COMPACT, &json_len);
-    ZITI_LOG(TRACE, "resp[%d,len=%zd] = %.*s",
-            result->success, json_len, (int)json_len, json);
-
-    if (result->data != NULL) {
-        tunnel_command tnl_res_cmd = {0};
-        if (parse_tunnel_command(&tnl_res_cmd, result->data, strlen(result->data)) >= 0) {
-            switch (tnl_res_cmd.command) {
-                case TunnelCommand_RemoveIdentity: {
-                    tunnel_identity_id tnl_delete_id;
-                    if (tnl_res_cmd.data != NULL && parse_tunnel_identity_id(&tnl_delete_id, tnl_res_cmd.data, strlen(tnl_res_cmd.data)) >= 0) {
-                        if (tnl_delete_id.identifier == NULL) {
-                            ZITI_LOG(ERROR, "Identity filename is not found in the remove identity request, not deleting the identity file");
-                            break;
-                        }
-#if _WIN32
-                        tunnel_identity *id = find_tunnel_identity(tnl_delete_id.identifier);
-                        if(id != NULL) {
-                            if (id->Services) {
-                                model_map hostnamesToRemove = {0};
-                                for (int index = 0; id->Services[index]; index++) {
-                                    tunnel_service *tnl_svc = id->Services[index];
-                                    if (tnl_svc->Addresses != NULL) {
-                                        for (int i = 0; tnl_svc->Addresses[i]; i++) {
-                                            tunnel_address *addr = tnl_svc->Addresses[i];
-                                            if (addr->IsHost &&
-                                                model_map_get(&hostnamesToRemove, addr->HostName) == NULL) {
-                                                model_map_set(&hostnamesToRemove, addr->HostName, "TRUE");
-                                            }
-                                        }
-                                    }
-                                }
-
-                                if (model_map_size(&hostnamesToRemove) > 0) {
-                                    remove_nrpt_rules(global_loop_ref, &hostnamesToRemove);
-                                }
-                            }
-                        } else {
-                            ZITI_LOG(WARN, "asked to remove identity, but identity was not found: %s", tnl_delete_id.identifier);
-                        }
-#endif
-                        delete_identity_from_instance(tnl_delete_id.identifier);
-                        free_tunnel_identity_id(&tnl_delete_id);
-                        // should be the last line in this function as it calls the mutex/lock
-                        save_tunnel_status_to_file();
-                    }
-
-                    break;
-                }
-                case TunnelCommand_IdentityOnOff: {
-                    if (!result->success) {
-                        break;
-                    }
-                    tunnel_on_off_identity on_off_id;
-                    if (tnl_res_cmd.data == NULL || parse_tunnel_on_off_identity(&on_off_id, tnl_res_cmd.data, strlen(tnl_res_cmd.data)) < 0) {
-                        free_tunnel_on_off_identity(&on_off_id);
-                        break;
-                    }
-                    set_ziti_status(on_off_id.onOff, on_off_id.identifier);
-                    // should be the last line in this function as it calls the mutex/lock
-                    save_tunnel_status_to_file();
-                    break;
-                }
-                case TunnelCommand_Unknown: {
-                    break;
-                }
-            }
-        }
-        free_tunnel_command(&tnl_res_cmd);
-    }
-
-    if (uv_is_active((const uv_handle_t *) ctx)) {
-        uv_buf_t buf;
-        size_t data_len = json_len + strlen("\n") + 1;
-        buf.base = calloc(data_len, sizeof(char));
-        snprintf(buf.base, data_len, "%s\n", json);
-        buf.len = strlen(buf.base);
-        uv_write_t *wr = calloc(1, sizeof(uv_write_t));
-        wr->data = buf.base;
-        uv_write(wr, (uv_stream_t *) ctx, &buf, 1, on_cmd_write);
-    }
-    free(json);
-}
-
-void tunnel_enroll_cb(const ziti_config *cfg, int status, const char *err, void *ctx) {
-    struct add_identity_request_s *add_id_req = ctx;
-
-    tunnel_result result = {
-            .success = false,
-            .error = NULL,
-            .data = NULL,
-            .code = IPC_ERROR,
-    };
-
-    if (status != ZITI_OK) {
-        ZITI_LOG(ERROR, "enrollment failed: %s(%d)", err, status);
-        result.error = "enrollment failed";
-        add_id_req->cmd_cb(&result, add_id_req->cmd_ctx);
-        free(add_id_req);
-        return;
-    }
-
-    FILE *f = add_id_req->add_id_ctx;
-
-    size_t len;
-    char *cfg_json = ziti_config_to_json(cfg, 0, &len);
-
-    if (fwrite(cfg_json, 1, len, f) != len) {
-        ZITI_LOG(ERROR, "failed to write config file");
-        fclose(f);
-        result.error = "failed to write config file";
-        add_id_req->cmd_cb(&result,  add_id_req->cmd_ctx);
-        free(add_id_req);
-        return;
-    }
-
-    free(cfg_json);
-    fflush(f);
-    fclose(f);
-
-    create_or_get_tunnel_identity(add_id_req->identifier, add_id_req->identifier_file_name);
-
-    // send load identity command to the controller
-    tunnel_load_identity load_identity_options = {
-            .identifier = add_id_req->identifier,
-            .path = add_id_req->identifier,
-            .apiPageSize = get_api_page_size(),
-    };
-    size_t json_len;
-    tunnel_command tnl_cmd = {
-            .command = TunnelCommand_LoadIdentity,
-            .data = tunnel_load_identity_to_json(&load_identity_options, MODEL_JSON_COMPACT, &json_len),
-    };
-    send_tunnel_command(&tnl_cmd, add_id_req->cmd_ctx);
-    free_tunnel_command(&tnl_cmd);
-    free(add_id_req);
-    save_tunnel_status_to_file();
-}
-
-static void enroll_ziti_async(uv_loop_t *loop, void *arg) {
-    struct add_identity_request_s *add_id_req = arg;
-
-    ziti_enroll_opts enroll_opts = {0};
-    enroll_opts.enroll_name = add_id_req->identifier;
-    enroll_opts.jwt_content = add_id_req->jwt_content;
-    enroll_opts.use_keychain = add_id_req->use_keychain;
-
-    ziti_enroll(&enroll_opts, loop, tunnel_enroll_cb, add_id_req);
-}
-
-static bool process_tunnel_commands(const tunnel_command *tnl_cmd, command_cb cb, void *ctx) {
-    tunnel_result result = {
-            .success = false,
-            .error = NULL,
-            .data = NULL,
-            .code = IPC_ERROR,
-    };
-    bool cmd_accepted = false;
-    switch (tnl_cmd->command) {
-        case TunnelCommand_SetLogLevel: {
-            cmd_accepted = true;
-
-            tunnel_set_log_level tunnel_set_log_level_cmd = {0};
-            if (tnl_cmd->data == NULL ||
-                parse_tunnel_set_log_level(&tunnel_set_log_level_cmd, tnl_cmd->data, strlen(tnl_cmd->data)) < 0 ||
-                tunnel_set_log_level_cmd.loglevel == NULL) {
-                result.error = "invalid command";
-                result.success = false;
-                break;
-            }
-
-            if (strcasecmp(ziti_log_level_label(), tunnel_set_log_level_cmd.loglevel) != 0) {
-                ziti_log_set_level_by_label(tunnel_set_log_level_cmd.loglevel);
-                ziti_tunnel_set_log_level(get_log_level(tunnel_set_log_level_cmd.loglevel));
-                const char *level = ziti_log_level_label();
-                set_log_level(level);
-                ZITI_LOG(INFO, "Log level is set to %s", level);
-            } else {
-                ZITI_LOG(INFO, "Log level is already set to %s", tunnel_set_log_level_cmd.loglevel);
-            }
-            result.success = true;
-            result.code = IPC_SUCCESS;
-
-            break;
-        }
-        case TunnelCommand_UpdateTunIpv4: {
-            cmd_accepted = true;
-
-            tunnel_tun_ip_v4 tunnel_tun_ip_v4_cmd = {0};
-            if (tnl_cmd->data == NULL || parse_tunnel_tun_ip_v4(&tunnel_tun_ip_v4_cmd, tnl_cmd->data, strlen(tnl_cmd->data)) < 0) {
-                result.error = "invalid command";
-                result.success = false;
-                free_tunnel_tun_ip_v4(&tunnel_tun_ip_v4_cmd);
-                break;
-            }
-            if (tunnel_tun_ip_v4_cmd.prefixLength < MINTUNPREFIXLENGTH || tunnel_tun_ip_v4_cmd.prefixLength > MAXTUNPREFIXLENGTH) {
-                result.error = "prefix length should be between 10 and 18";
-                result.success = false;
-                break;
-            }
-            char* tun_ip_str = strdup(tunnel_tun_ip_v4_cmd.tunIP);
-            // make a copy, so we can free it later - validating ip address input
-            char* tun_ip_cpy = tun_ip_str;
-            char* ip_ptr = strtok(tun_ip_str, "."); //cut the string using dot delimiter
-            if (ip_ptr == NULL) {
-                result.error = "Invalid IP address";
-                result.success = false;
-                break;
-            }
-            int dots = 0;
-            bool validationStatus = true;
-            while (ip_ptr) {
-                bool isInt = true;
-                char* ip_str = ip_ptr;
-                while (*ip_str) {
-                    if(!isdigit(*ip_str)){ //if the character is not a number, break
-                        isInt = false;
-                        validationStatus = false;
-                        break;
-                    }
-                    ip_str++; //point to next character
-                }
-                if (!isInt) {
-                    break;
-                }
-                int num = atoi(ip_ptr); //convert substring to number
-                if (num >= 0 && num <= 255) {
-                    ip_ptr = strtok(NULL, "."); //cut the next part of the string
-                    if (ip_ptr != NULL)
-                        dots++; //increase the dot count
-                } else {
-                    validationStatus = false;
-                    break;
-                }
-            }
-            free(tun_ip_cpy);
-            if (dots != 3 || !validationStatus) {
-                result.error = "Invalid IP address";
-                result.success = false;
-                free_tunnel_tun_ip_v4(&tunnel_tun_ip_v4_cmd);
-                break;
-            }
-            if (tunnel_tun_ip_v4_cmd.tunIP == NULL) {
-                result.error = "Tun IP is null";
-                result.success = false;
-                free_tunnel_tun_ip_v4(&tunnel_tun_ip_v4_cmd);
-                break;
-            }
-            set_tun_ipv4_into_instance(tunnel_tun_ip_v4_cmd.tunIP,
-                                       (int)tunnel_tun_ip_v4_cmd.prefixLength,
-                                       tunnel_tun_ip_v4_cmd.addDns);
-            result.success = true;
-            result.code = IPC_SUCCESS;
-            break;
-        }
-        case TunnelCommand_Status: {
-            cmd_accepted = true;
-            tunnel_status* status = get_tunnel_status();
-            result.success = true;
-            result.code = IPC_SUCCESS;
-            size_t json_len;
-            result.data = tunnel_status_to_json(status, MODEL_JSON_COMPACT, &json_len);
-            break;
-        }
-
-        case TunnelCommand_AddIdentity : {
-            cmd_accepted = true;
-            tunnel_add_identity tunnel_add_identity_cmd = {0};
-            if (tnl_cmd->data == NULL ||
-                parse_tunnel_add_identity(&tunnel_add_identity_cmd, tnl_cmd->data, strlen(tnl_cmd->data)) < 0 ||
-                (tunnel_add_identity_cmd.jwtFileName == NULL && tunnel_add_identity_cmd.jwtContent == NULL)) {
-                result.error = "invalid command";
-                result.success = false;
-                free_tunnel_add_identity(&tunnel_add_identity_cmd);
-                break;
-            }
-
-            if (tunnel_add_identity_cmd.jwtFileName == NULL) {
-                result.error = "identity filename not provided";
-                result.success = false;
-                break;
-            }
-
-            if (tunnel_add_identity_cmd.jwtContent == NULL) {
-                result.error = "jwt content not provided";
-                result.success = false;
-                break;
-            }
-
-            if (config_dir == NULL) {
-                result.error = "config directory not set";
-                result.success = false;
-                break;
-            }
-
-            char* extension = strstr(tunnel_add_identity_cmd.jwtFileName, ".jwt");
-            size_t length;
-            if (extension != NULL) {
-                length = extension - tunnel_add_identity_cmd.jwtFileName;
-            } else {
-                length = strlen(tunnel_add_identity_cmd.jwtFileName);
-            }
-            char new_identifier[FILENAME_MAX] = {0};
-            char new_identifier_name[FILENAME_MAX] = {0};
-            if ((strlen(config_dir) + length + 6) >  FILENAME_MAX - 1 ) {
-                ZITI_LOG(ERROR, "failed to create file %s%c%s.json, The length of the file name is longer than %d", config_dir, PATH_SEP, tunnel_add_identity_cmd.jwtFileName, FILENAME_MAX);
-                result.error = "invalid file name";
-                result.success = false;
-                free_tunnel_add_identity(&tunnel_add_identity_cmd);
-                break;
-            }
-            strncpy(new_identifier_name, tunnel_add_identity_cmd.jwtFileName, length);
-            snprintf(new_identifier, FILENAME_MAX, "%s%c%s.json", config_dir, PATH_SEP, new_identifier_name);
-            FILE *outfile;
-            if ((outfile = fopen(new_identifier, "wb")) == NULL) {
-                ZITI_LOG(ERROR, "failed to open file %s: %s(%d)", new_identifier, strerror(errno), errno);
-                result.error = "invalid file name";
-                result.success = false;
-                free_tunnel_add_identity(&tunnel_add_identity_cmd);
-                break;
-            }
-
-            struct add_identity_request_s *add_id_req = calloc(1, sizeof(struct add_identity_request_s));
-            add_id_req->cmd_ctx = ctx;
-            add_id_req->cmd_cb = cb;
-            add_id_req->add_id_ctx = outfile;
-            add_id_req->identifier = strdup(new_identifier);
-            add_id_req->identifier_file_name = strdup(new_identifier_name);
-            add_id_req->jwt_content = strdup(tunnel_add_identity_cmd.jwtContent);
-            add_id_req->use_keychain = tunnel_add_identity_cmd.useKeychain;
-
-            enroll_ziti_async(global_loop_ref, add_id_req);
-            free_tunnel_add_identity(&tunnel_add_identity_cmd);
-            return true;
-        }
-#if _WIN32
-        case TunnelCommand_ServiceControl: {
-            cmd_accepted = true;
-            tunnel_service_control tunnel_service_control_opts = {0};
-            if (tnl_cmd->data == NULL ||
-                parse_tunnel_service_control(&tunnel_service_control_opts, tnl_cmd->data, strlen(tnl_cmd->data)) < 0) {
-                result.error = "invalid command";
-                result.success = false;
-                free_tunnel_service_control(&tunnel_service_control_opts);
-                break;
-            }
-            result.success = true;
-            result.code = IPC_SUCCESS;
-            if (tunnel_service_control_opts.operation != NULL && strcmp(tunnel_service_control_opts.operation, "stop") == 0) {
-                // stops the windows service in scm
-                if (!stop_windows_service()) {
-                    ZITI_LOG(INFO, "Could not send stop signal to scm, Tunnel must not be started as service");
-                    stop_tunnel_and_cleanup();
-                    uv_stop(global_loop_ref);
-                }
-            }
-            free_tunnel_service_control(&tunnel_service_control_opts);
-            break;
-        }
-        case TunnelCommand_StatusChange: {
-            cmd_accepted = true;
-            tunnel_status_change tunnel_status_change_opts = {0};
-            if (tnl_cmd->data == NULL ||
-                parse_tunnel_status_change(&tunnel_status_change_opts, tnl_cmd->data, strlen(tnl_cmd->data)) < 0) {
-                result.error = "invalid command";
-                result.success = false;
-                free_tunnel_status_change(&tunnel_status_change_opts);
-                break;
-            }
-            result.success = true;
-            result.code = IPC_SUCCESS;
-
-            endpoint_status_change(tunnel_status_change_opts.woken, tunnel_status_change_opts.unlocked);
-
-            free_tunnel_status_change(&tunnel_status_change_opts);
-        }
-#endif
-    }
-    if (cmd_accepted) {
-        cb(&result, ctx);
-        if (result.success) {
-            // should be the last line in this function as it calls the mutex/lock
-            save_tunnel_status_to_file();
-        }
-        if (result.data) {
-            free(result.data);
-        }
-        return true;
-    } else {
-        return false;
-    }
-}
-
-static void process_ipc_command(uv_stream_t *s, json_object *json) {
-    tunnel_command tnl_cmd = {0};
-    if (tunnel_command_from_json(&tnl_cmd, json) >= 0) {
-        // process_tunnel_commands is used to update the log level and the tun ip information in the config file through IPC command.
-        // So when the user restarts the tunnel, the new values will be taken.
-        // The config file can be modified only from ziti-edge-tunnel.c file.
-        int status = process_tunnel_commands(&tnl_cmd, on_command_resp, s);
-        if (!status) {
-            // process_cmd will delegate the requests to ziti_tunnel_ctrl.c , which is used to perform the operations against the controller.
-            // config.file cannot be modified or read from that class
-            CMD_CTRL->process(&tnl_cmd, on_command_resp, s);
-        }
-    } else {
-        tunnel_result resp = {
-                .success = false,
-                .error = "failed to parse command",
-                .code = IPC_ERROR,
-        };
-        on_command_resp(&resp, s);
-    }
-    free_tunnel_command(&tnl_cmd);
-}
-
-
-static void on_cmd(uv_stream_t *s, ssize_t len, const uv_buf_t *b) {
-    struct ipc_conn_s *ipc_client = (struct ipc_conn_s*)s;
-    if (len < 0) {
-        if (len != UV_EOF) {
-            ZITI_LOG(WARN, "received from client - %s. Closing connection.", uv_err_name(len));
-        }
-
-        LIST_REMOVE(ipc_client, _next_ipc_cmd);
-
-        json_tokener *tokener = s->data;
-        if (tokener) json_tokener_free(tokener);
-        uv_close((uv_handle_t *) &ipc_client->ipc, (uv_close_cb) free);
-        ZITI_LOG(DEBUG, "IPC client connection closed, count: %d", sizeof_ipc_clients_list());
-
-    } else {
-        ZITI_LOG(DEBUG, "received cmd <%.*s>", (int) len, b->base);
-
-        json_tokener *parser = s->data;
-
-        size_t processed = 0;
-        while (processed < len) {
-            json_object *json = json_tokener_parse_ex(parser, b->base + processed, (int) (len - processed));
-            size_t end = json_tokener_get_parse_end(parser);
-            processed += end;
-            if (json) {
-                process_ipc_command(s, json);
-                json_object_put(json);
-            } else if (json_tokener_get_error(parser) != json_tokener_continue) {
-                ZITI_LOG(ERROR, "failed to parse json command: %s, received[%.*s]",
-                         json_tokener_error_desc(json_tokener_get_error(parser)),
-                         (int) len, b->base);
-                json_tokener_free(parser);
-                LIST_REMOVE(ipc_client, _next_ipc_cmd);
-                uv_close((uv_handle_t *) &ipc_client->ipc, (uv_close_cb) free);
-                break;
-            }
-        }
-    }
-
-    free(b->base);
-}
-
-static void on_cmd_client(uv_stream_t *s, int status) {
-    int current_ipc_channels = sizeof_ipc_clients_list();
-    struct ipc_conn_s *cmd_conn = calloc(1, sizeof(struct ipc_conn_s));
-    cmd_conn->ipc.data = json_tokener_new();
-    uv_pipe_init(s->loop, &cmd_conn->ipc, 0);
-    uv_accept(s, (uv_stream_t *) &cmd_conn->ipc);
-    uv_read_start((uv_stream_t *) &cmd_conn->ipc, cmd_alloc, on_cmd);
-    LIST_INSERT_HEAD(&ipc_clients_list, cmd_conn, _next_ipc_cmd);
-    ZITI_LOG(DEBUG,"Received IPC client connection request, count: %d", ++current_ipc_channels);
-}
-
-static int start_cmd_socket(uv_loop_t *l) {
-
-    if (uv_is_active((const uv_handle_t *) &cmd_server)) {
-        return 0;
-    }
-
-    uv_fs_t fs;
-    uv_fs_unlink(l, &fs, sockfile, NULL);
-
-#define CHECK_UV(op) do{ \
-    int uv_rc = (op);    \
-    if (uv_rc != 0) {    \
-       ZITI_LOG(WARN, "failed to open IPC socket op=[%s] err=%d[%s]", #op, uv_rc, uv_strerror(uv_rc));\
-       goto uv_err; \
-    }                    \
-    } while(0)
-
-
-    CHECK_UV(uv_pipe_init(l, &cmd_server, 0));
-    CHECK_UV(uv_pipe_bind(&cmd_server, sockfile));
-    CHECK_UV(uv_pipe_chmod(&cmd_server, UV_WRITABLE | UV_READABLE));
-
-    uv_unref((uv_handle_t *) &cmd_server);
-
-    CHECK_UV(uv_listen((uv_stream_t *) &cmd_server, 0, on_cmd_client));
-
-    return 0;
-
-    uv_err:
-    return -1;
-}
-
-static void on_events_client(uv_stream_t *s, int status) {
-    int current_events_channels = sizeof_event_clients_list();
-    uv_pipe_t* event_conn = malloc(sizeof(uv_pipe_t));
-    uv_pipe_init(s->loop, event_conn, 0);
-    uv_accept(s, (uv_stream_t *) event_conn);
-    struct event_conn_s *event_client_conn = calloc(1, sizeof(struct event_conn_s));
-    event_client_conn->event_client_conn = event_conn;
-    LIST_INSERT_HEAD(&event_clients_list, event_client_conn, _next_event);
-    ZITI_LOG(DEBUG,"Received events client connection request, count: %d", ++current_events_channels);
-
-    // send status message immediately
-    send_tunnel_status("status");
-}
-
-
-void on_write_event(uv_write_t* req, int status) {
-    if (status < 0) {
-        ZITI_LOG(ERROR,"Could not sent events message. Write error %s\n", uv_err_name(status));
-        if (status == UV_EPIPE) {
-            struct event_conn_s *event_client;
-            LIST_FOREACH(event_client, &event_clients_list, _next_event) {
-                if (event_client->event_client_conn == (uv_pipe_t*) req->handle) {
-                    break;
-                }
-            }
-            if (event_client) {
-                uv_close((uv_handle_t *) event_client->event_client_conn, (uv_close_cb) free);
-                event_client->event_client_conn = NULL;
-                int current_event_connection_count = sizeof_event_clients_list();
-                ZITI_LOG(WARN,"Events client connection closed, count : %d", current_event_connection_count);
-
-            }
-
-        }
-    } else {
-        ZITI_LOG(TRACE,"Events message is sent.");
-    }
-    if (req->data) {
-        free(req->data);
-    }
-    free(req);
-}
-
-static void send_events_message(const void *message, to_json_fn to_json_f, bool displayEvent) {
-    size_t data_len = 0;
-    char *json = to_json_f(message, MODEL_JSON_COMPACT, &data_len);
-    if (json == NULL) {
-        ZITI_LOG(ERROR, "failed to serialize event");
-        return;
-    }
-    if (displayEvent) {
-        ZITI_LOG(DEBUG,"Events Message => %s", json);
-    }
-
-    if (!LIST_EMPTY(&event_clients_list)) {
-        struct event_conn_s *event_client;
-        int events_deleted = 0;
-        LIST_FOREACH(event_client, &event_clients_list, _next_event) {
-            int err = 0;
-            if (event_client->event_client_conn != NULL) {
-                uv_buf_t buf;
-                data_len = data_len + strlen("\n") + 1;
-                buf.base = calloc(data_len, sizeof(char));
-                snprintf(buf.base, data_len, "%s\n", json);
-                buf.len = strlen(buf.base);
-                uv_write_t *wr = calloc(1, sizeof(uv_write_t));
-                wr->data = buf.base;
-                err = uv_write(wr, (uv_stream_t *)event_client->event_client_conn, &buf, 1, on_write_event);
-            }
-            if (err < 0){
-                ZITI_LOG(ERROR,"Events client write operation failed, received error - %s", uv_err_name(err));
-                if (err == UV_EPIPE) {
-                    uv_close((uv_handle_t *) event_client->event_client_conn, (uv_close_cb) free);
-                    event_client->event_client_conn = NULL;
-                    events_deleted++;
-                    ZITI_LOG(WARN,"Events client connection closed");
-                }
-            }
-        }
-        if (events_deleted > 0) {
-            int current_event_connection_count = sizeof_event_clients_list();
-            ZITI_LOG(WARN,"Events client connection current count : %d", current_event_connection_count);
-        }
-
-    }
-    free(json);
-}
-
-static int start_event_socket(uv_loop_t *l) {
-
-    if (uv_is_active((const uv_handle_t *) &event_server)) {
-        return 0;
-    }
-
-    uv_fs_t fs;
-    uv_fs_unlink(l, &fs, eventsockfile, NULL);
-
-    CHECK_UV(uv_pipe_init(l, &event_server, 0));
-    CHECK_UV(uv_pipe_bind(&event_server, eventsockfile));
-    CHECK_UV(uv_pipe_chmod(&event_server, UV_WRITABLE | UV_READABLE));
-
-    uv_unref((uv_handle_t *) &event_server);
-
-    CHECK_UV(uv_listen((uv_stream_t *) &event_server, 0, on_events_client));
-
-    return 0;
-
-    uv_err:
-    return -1;
-}
-
-
-static void tnl_transfer_rates(const tunnel_identity_metrics *metrics, void *ctx) {
-    tunnel_identity *tnl_id = ctx;
-    if (metrics->up != NULL) {
-        tnl_id->Metrics.Up = (int) strtol(metrics->up, NULL, 10);
-    }
-    if (metrics->down != NULL) {
-        tnl_id->Metrics.Down = (int) strtol(metrics->down, NULL, 10);
-    }
-}
-
-static void on_command_inline_resp(const tunnel_result* result, void *ctx) {
-    tunnel_command_inline *tnl_cmd_inline = ctx;
-
-    if (tnl_cmd_inline == NULL) {
-        return;
-    }
-
-    if (result->data != NULL && strlen(result->data) > 0) {
-        switch (tnl_cmd_inline->command) {
-            case TunnelCommand_GetMetrics: {
-                if (result->success) {
-                    tunnel_identity_metrics id_metrics = {0};
-                    if (parse_tunnel_identity_metrics(&id_metrics, result->data, strlen(result->data)) < 0) {
-                        ZITI_LOG(ERROR, "Could not fetch metrics data");
-                    } else {
-                        tunnel_identity *tnl_id = find_tunnel_identity(tnl_cmd_inline->identifier);
-                        tnl_transfer_rates(&id_metrics, tnl_id);
-                    }
-                    free_tunnel_identity_metrics(&id_metrics);
-                }
-                break;
-            }
-            default: {
-                ZITI_LOG(ERROR, "Tunnel command not supported %d", tnl_cmd_inline->command);
-            }
-        }
-    }
-
-    if (tnl_cmd_inline != NULL) {
-        free_tunnel_command_inline(tnl_cmd_inline);
-        free(tnl_cmd_inline);
-    }
-}
-
-static void send_tunnel_command(const tunnel_command *tnl_cmd, void *ctx) {
-    CMD_CTRL->process(tnl_cmd, on_command_resp, ctx);
-}
-
-static void send_tunnel_command_inline(const tunnel_command *tnl_cmd, void *ctx) {
-    CMD_CTRL->process(tnl_cmd, on_command_inline_resp, ctx);
 }
 
 static char* addUnit(int count, char* unit) {
@@ -1732,8 +992,8 @@ static void run_tunneler_loop(uv_loop_t* ziti_loop) {
     int rc0 = 0, rc1;
     rc0 = rc1 = make_socket_path(ziti_loop);
     if (rc0 == 0) {
-        rc0 = start_cmd_socket(ziti_loop);
-        rc1 = start_event_socket(ziti_loop);
+        rc0 = start_cmd_socket(ziti_loop, sockfile);
+        rc1 = start_event_socket(ziti_loop, eventsockfile);
     }
 
     if (rc0 < 0 || rc1 < 0) {
@@ -3358,16 +2618,3 @@ int main(int argc, char *argv[]) {
     commandline_run(&main_cmd, argc, argv);
     return 0;
 }
-
-// ******* TUNNEL EVENT BROADCAST MESSAGES
-IMPL_ENUM(event_severity, EVENT_SEVERITY)
-IMPL_MODEL(status_event, STATUS_EVENT)
-IMPL_MODEL(action_event, ACTION_EVENT)
-IMPL_MODEL(identity_event, IDENTITY_EVENT)
-IMPL_MODEL(services_event, SERVICES_EVENT)
-IMPL_MODEL(tunnel_status_event, TUNNEL_STATUS_EVENT)
-IMPL_MODEL(mfa_status_event, MFA_STATUS_EVENT)
-IMPL_MODEL(tunnel_metrics_event, TUNNEL_METRICS_EVENT)
-IMPL_MODEL(notification_message, TUNNEL_NOTIFICATION_MESSAGE)
-IMPL_MODEL(notification_event, TUNNEL_NOTIFICATION_EVENT)
-
